- changed router setup
